refactor: FloorMapCanvas.tsxの関数や変数などの処理の部分をカスタムフックとして切り出した。
